edit bottomsheet edit TopBar add new Flows and Observers add DataStoreRepository edit ApiService

+++ app/src/main/java/com/dev_marinov/kinopoiskapp/data/rating/local/RatingEntity.kt
@@ -22,7 +22,7 @@ data class RatingEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Int,
     val await: Double,
-    val filmCritics: Int,
+    val filmCritics: Double,
     val imdb: Double,
     val kp: Double,
     val russianFilmCritics: Double,
