+++ app/src/main/java/com/dev_marinov/kinopoiskapp/domain/model/Rating.kt
@@ -4,7 +4,7 @@ data class Rating(
     val id: Int, // in new api this value not found
     val kp: Double,
     val imdb: Double,
-    val filmCritics: Int,
+    val filmCritics: Double,
     val russianFilmCritics: Double,
     val await: Double,
     val movieId: Int?
